+++ comics/comics/manalanextdoor.py
@@ -1,4 +1,4 @@
-from comics.aggregator.crawler import CrawlerBase, CrawlerImage
+from comics.aggregator.crawler import CrawlerBase
 from comics.core.comic_data import ComicDataBase
 
 
@@ -7,17 +7,11 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://www.manalanextdoor.com/'
     start_date = '2011-01-23'
+    end_date = '2012-11-14'
+    active = False
     rights = 'Humon'
 
 
 class Crawler(CrawlerBase):
-    time_zone = 'Europe/Copenhagen'
-
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://feeds.feedburner.com/manalanextdoor')
-        for entry in feed.all():
-            url = entry.summary.src('img[src*="/art/"]')
-            if url is not None:
-                url = url.replace('/thumb', '')
-            title = entry.title
-            return CrawlerImage(url, title)
+        pass  # Comic no longer published